Unified rebuildRows.
