+++ uui-core/src/data/processing/views/Tree.ts
@@ -307,7 +307,6 @@ export class Tree<TItem, TId> {
             if (selectedId != null) {
                 selectedIdsMap.set(selectedId, true);
             }
-
             if (options.cascade) {
                 // check all children recursively
                 forEachChildren(id => selectedIdsMap.set(id, true));
